Modified GUI layout and added warning functionality

+++ src/application/tuple.java
@@ -3,14 +3,14 @@ package application;
 //This is a basic class that stores a tuple (x, y) point
 
 public class tuple {
-	public float x;
-	public float y;
-	public tuple(float _x, float _y) {
+	public double x;
+	public double y;
+	public tuple(double _x, double _y) {
 		x = _x;
 		y = _y;
 	}
 	@Override
 	public String toString() {
-		return "(" + Float.toString(x) + ", " + Float.toString(y) + ")";
+		return "(" + Double.toString(x) + ", " + Double.toString(y) + ")";
 	}
 }
